implemented isBroadcastStreamAllowedForIP

+++ src/com/hts/service/IpAddressServiceImpl.java
@@ -7,7 +7,10 @@ import org.apache.log4j.Logger;
 
 import com.hts.dao.DAO;
 import com.hts.dao.IpAddressDAOHibernateImpl;
+import com.hts.entities.Channel;
 import com.hts.entities.IpAddress;
+import com.hts.entities.Room;
+import com.hts.entities.SubscriptionPackage;
 import com.hts.exceptions.AppException;
 
 public class IpAddressServiceImpl implements IIpAddressService {
@@ -24,15 +27,18 @@ public class IpAddressServiceImpl implements IIpAddressService {
 
 	@Override
 	public void delete(IpAddress ipAddress) throws AppException {
-		ipAddressDAO.delete(ipAddress);
-		DAO.close();
-		log.info("deleted ipAddress: " + ipAddress.getIpAddress());
+		if (ipAddress != null) {
+
+			ipAddressDAO.delete(ipAddress);
+			DAO.close();
+			log.info("deleted ipAddress: " + ipAddress.getIpAddress());
+		}
 
 	}
 
 	@Override
-	public List<IpAddress> getByName(String name) throws AppException {
-		return ipAddressDAO.getByName(name);
+	public List<IpAddress> getByIp(String ipAddress) throws AppException {
+		return ipAddressDAO.getByIp(ipAddress);
 	}
 
 	@Override
@@ -41,62 +47,42 @@ public class IpAddressServiceImpl implements IIpAddressService {
 	}
 
 	@Override
-	public IpAddress create(String name) throws UnknownHostException, AppException {
-		
+	public IpAddress create(String name) throws UnknownHostException,
+			AppException {
+
 		IpAddress ipAddress = new IpAddress(name);
-		
+
 		ipAddressDAO.create(ipAddress);
 		DAO.close();
 		log.info("updated ipAddress: " + ipAddress.getIpAddress());
 		return ipAddress;
-	
+
 	}
 
-	// public IpAddress create(String name) throws AppException {
-	//
-	// Hotel hotel = hotelDAO.getAll().get(0);
-	//
-	// IpAddress r = ipAddressDAO.create(new IpAddress(name, hotel));
-	// DAO.close();
-	// log.info("created ipAddress: " + r);
-	// return r;
-	// }
-	//
-	// @Override
-	// public IpAddress create(String name, Hotel hotel) throws AppException {
-	//
-	// if (hotel == null)
-	// throw new
-	// AppException("Hotel must not be null. Use IpAddressServiceImpl.create(String str).");
-	//
-	// IpAddress r = ipAddressDAO.create(new IpAddress(name, hotel));
-	// DAO.close();
-	// log.info("created ipAddress: " + r);
-	// return r;
-	// }
-	//
-	// @Override
-	// public void update(IpAddress ipAddress) throws AppException {
-	// ipAddressDAO.save(ipAddress);
-	// DAO.close();
-	// log.info("updated ipAddress: " + ipAddress);
-	// }
-	//
-	// @Override
-	// public void delete(IpAddress ipAddress) throws AppException {
-	// ipAddressDAO.delete(ipAddress);
-	// DAO.close();
-	// log.info("deleted ipAddress: " + ipAddress);
-	// }
-	//
-	// @Override
-	// public List<IpAddress> getByName(String name) throws AppException {
-	// return ipAddressDAO.getByName(name);
-	// }
-	//
-	// @Override
-	// public IpAddress getById(Integer id) throws AppException {
-	// return ipAddressDAO.getById(id);
-	// }
+	@Override
+	public boolean isBroadcastStreamAllowedForIP(String ipAddress,
+			String broadcastStreamName) throws UnknownHostException,
+			AppException {
+
+		List<IpAddress> ipAddresses = ipAddressDAO.getByIp(ipAddress);
+		if (ipAddresses.size()==0)
+			return false;
+		
+		IpAddress ipAddr = ipAddresses.get(0);
+		
+		Room room = ipAddr.getRoom();
+		if (room==null)
+			return false;
+		SubscriptionPackage subscriptionPackage = room.getSubscriptionPackage();
+		if (subscriptionPackage ==null)
+			return false;
+		
+		List<Channel> chanels = (List<Channel>) subscriptionPackage
+				.getChannels();
+		Channel channel = new ChannelServiceImpl()
+				.getByBroadcastName(broadcastStreamName);
+		return chanels.contains(channel);
+
+	}
 
 }
